[ch #163774621] add parties data stuctures on base test class
