+++ tests/test_base.py
@@ -6,16 +6,42 @@ from app import create_app
 class BaseTest(unittest.TestCase):
     """define setup and teardown methods."""
     def setUp(self):
-        app = create_app()
+        app = create_app("testing")
         self.client = app.test_client()
 
         #a dictionary containing a party
-        self.party = {
-                "name" : "ANC",
-                "hqAddress" : "Kakamega",
-                "logoUrl" : "https://goo.gl/images/B9U4PK",
-            }
+        self.party_with_data = {
+            "name" : "ANC",
+            "hqAddress" : "Kakamega",
+            "logoUrl" : "https://goo.gl/images/B9U4PK",
+        }
 
+        #a party with empty party name
+        self.party_with_name_missing = {
+            "name" : " ",
+            "hqAddress" : "Kakamega",
+            "logoUrl" : "https://goo.gl/images/B9U4PK",
+        }
+
+        #a party with empty party hq
+        self.party_with_hq_missing = {
+            "name" : "Jubilee",
+            "hqAddress" : "Muthaiga",
+            "logoUrl" : "https://goo.gl/images/7hU72H",
+        }
+
+        #a party with empty fields
+        self.party_with_empty_fields = {
+            "name" : " ",
+            "hqAddress" : " ",
+            "logoUrl" : " ",
+        }
+        #a party with a name of wrong data type
+        self.party_with_name_of_wrong_data_type = {
+            "name" : 4564,
+            "hqAddress" : "Machakos",
+            "logoUrl" : "https://goo.gl/images/3RKgQ6",
+        }
         #a list of containing dictionaries containing parties
         self.parties = [
             {
@@ -35,29 +61,6 @@ class BaseTest(unittest.TestCase):
             }
         ]
 
-    def test_add_party(self):
-        """ensure a new party can be added to the political party list."""
-        with self.client:
-            response = self.client.post(
-                '/api/v1/parties',
-                data = json.dumps(self.party),
-                content_type = "application/json",
-            )
-            data = json.loads(response.data)
-            self.assertEqual(response.status_code, 201)
-            self.assertIn('ANC was added', data['success'])
-
-    def test_get_parties(self):
-        """a list of political parties is returned."""
-        with self.client:
-            response = self.client.get(
-                '/api/v1/parties',
-                data = json.dumps(self.parties),
-                content_type = "application/json",
-            )
-            data = json.loads(response.data)
-            self.assertEqual(response.status_code, 200)
-            self.assertIsInstance(data['data'], list)
